software/python/requirements.txt
pretty_midi >= 0.2.11
matplotlib >= 3.10.8
pyyaml >= 6.0.3